Full conversion to zTemplate is now complete.

+++ lib/readmarker.php
@@ -1,7 +1,7 @@
 <?php
 /**
  * QSF Portal
- * Copyright (c) 2006-2015 The QSF Portal Development Team
+ * Copyright (c) 2006-2019 The QSF Portal Development Team
  * https://github.com/Arthmoor/QSF-Portal
  *
  * Based on:
@@ -22,8 +22,8 @@
  *
  **/
 
-if (!defined('QUICKSILVERFORUMS')) {
-	header('HTTP/1.0 403 Forbidden');
+if( !defined( 'QUICKSILVERFORUMS' ) ) {
+	header( 'HTTP/1.0 403 Forbidden' );
 	die;
 }
 
@@ -46,50 +46,51 @@ class readmarker extends forumutils
 	var $readmarkers = array();	// Data for user as pulled from the database
 	var $forumtopics = array();	// Cache of topics within forums
 	var $cleanupchance = false;	// Set to true if we want a cleanup operation done
-	
+
 	/**
 	 * Constructor. Initalise the read marker for guest (cookie and session)
 	 * or user (readmark table)
 	 *
 	 * @param $qsf - Quicksilver Forums module
 	 **/
-	function readmarker(&$qsf)
+	public function __construct( &$qsf )
 	{
-		parent::forumutils($qsf);
+		parent::__construct( $qsf );
 
 		$this->time = &$qsf->time;
 		$this->day_in_seconds = 86400;
-		
+
 		// To initalise ourselves we need to look at the user
-		if ($qsf->perms->is_guest) {
+		if( $qsf->perms->is_guest ) {
 			// With a guest user we can try and read/set a cookie but that's all!
-			if (isset($qsf->cookie[$this->sets['cookie_prefix'] . 'lastallread']) && $qsf->cookie[$this->sets['cookie_prefix'] . 'lastallread'] < ($this->time - ($this->day_in_seconds * 2))) {
-				$this->last_read_all = intval($qsf->cookie[$this->sets['cookie_prefix'] . 'lastallread']);
+			if( isset( $qsf->cookie[$this->sets['cookie_prefix'] . 'lastallread'] ) && $qsf->cookie[$this->sets['cookie_prefix'] . 'lastallread'] < ( $this->time - ( $this->day_in_seconds * 2 ) ) ) {
+				$this->last_read_all = intval( $qsf->cookie[$this->sets['cookie_prefix'] . 'lastallread'] );
 			} else {
 				$this->last_read_all = $this->time - $this->day_in_seconds;
 
-				setcookie($this->sets['cookie_prefix'] . 'lastallread', $this->last_read_all, $qsf->time + $this->sets['logintime'], $this->sets['cookie_path'], $this->sets['cookie_domain'], $this->sets['cookie_secure'], true );
+				setcookie( $this->sets['cookie_prefix'] . 'lastallread', $this->last_read_all, $qsf->time + $this->sets['logintime'], $this->sets['cookie_path'], $this->sets['cookie_domain'], $this->sets['cookie_secure'], true );
 			}
-			if (!isset($_SESSION[$this->sets['cookie_prefix'] . 'read_topics'])) {
+
+			if( !isset( $_SESSION[$this->sets['cookie_prefix'] . 'read_topics'] ) ) {
 				$_SESSION[$this->sets['cookie_prefix'] . 'read_topics'] = array();
 			}
 
 			$this->guest_mode = true;
 		} else {
 			// Get the user ID and the lastallread value and prepare to
-			if ($qsf->user['user_lastallread']) {
+			if( $qsf->user['user_lastallread'] ) {
 				$this->last_read_all = $qsf->user['user_lastallread'];
-			} elseif (isset($qsf->cookie[$this->sets['cookie_prefix'] . 'lastallread'])) {
-				$this->last_read_all = intval($qsf->cookie[$this->sets['cookie_prefix'] . 'lastallread']);
+			} elseif( isset( $qsf->cookie[$this->sets['cookie_prefix'] . 'lastallread'] ) ) {
+				$this->last_read_all = intval( $qsf->cookie[$this->sets['cookie_prefix'] . 'lastallread'] );
 			} else {
 				$this->last_read_all = $this->time - $this->day_in_seconds;
 			}
-			
+
 			$this->guest_mode = false;
-			$this->user_id = intval($qsf->user['user_id']);
+			$this->user_id = intval( $qsf->user['user_id'] );
 		}
 	}
-	
+
 	/**
 	 * A kind of deconstructor but not always run
 	 *
@@ -97,13 +98,13 @@ class readmarker extends forumutils
 	 * @author Geoffrey Dunn <geoff@warmage.com>
 	 * @since 1.2.0
 	 **/
-	function cleanup()
+	public function cleanup()
 	{
-		if ($this->cleanupchance && Rand(1,20) == 1) {
+		if( $this->cleanupchance && Rand( 1, 20 ) == 1 ) {
 			$this->_cleanup_readmarks();
 		}
 	}
-	
+
 	/**
 	 * Updates all records of all topics, marking them as read
 	 *
@@ -111,25 +112,26 @@ class readmarker extends forumutils
 	 * @author Geoffrey Dunn <geoff@warmage.com>
 	 * @since 1.2.0
 	 **/
-	function mark_all_read($time)
+	public function mark_all_read( $time )
 	{
-		$time = intval($time);
-		
-		if ($this->last_read_all >= $time) return; // Do Nothing
-		
+		$time = intval( $time );
+
+		if( $this->last_read_all >= $time )
+			return; // Do Nothing
+
 		$this->last_read_all = $time;
+
 		// Clean out unneeded entries
-		if ($this->guest_mode) {
-			foreach (array_keys($_SESSION[$this->sets['cookie_prefix'] . 'read_topics']) as $topic) {
-				if ($_SESSION[$this->sets['cookie_prefix'] . 'read_topics'][$topic] < $time) {
-					unset($_SESSION[$this->sets['cookie_prefix'] . 'read_topics'][$topic]);
+		if( $this->guest_mode ) {
+			foreach( array_keys( $_SESSION[$this->sets['cookie_prefix'] . 'read_topics'] ) as $topic ) {
+				if( $_SESSION[$this->sets['cookie_prefix'] . 'read_topics'][$topic] < $time ) {
+					unset( $_SESSION[$this->sets['cookie_prefix'] . 'read_topics'][$topic] );
 				}
 			}
-			setcookie($this->sets['cookie_prefix'] . 'lastallread', $time, $this->time + $this->sets['logintime'], $this->sets['cookie_path'], $this->sets['cookie_domain'], $this->sets['cookie_secure'], true );
+			setcookie( $this->sets['cookie_prefix'] . 'lastallread', $time, $this->time + $this->sets['logintime'], $this->sets['cookie_path'], $this->sets['cookie_domain'], $this->sets['cookie_secure'], true );
 		} else {
-			$this->db->query("UPDATE %pusers SET user_lastallread=%s WHERE user_id=%d", $time, $this->user_id);
-			$this->db->query("DELETE FROM %preadmarks 
-				WHERE readmark_user=%d AND readmark_lastread<%d", $this->user_id, $time);
+			$this->db->query( "UPDATE %pusers SET user_lastallread=%s WHERE user_id=%d", $time, $this->user_id );
+			$this->db->query("DELETE FROM %preadmarks WHERE readmark_user=%d AND readmark_lastread<%d", $this->user_id, $time );
 		}
 		$this->readmarkers_loaded = false;
 	}
@@ -142,17 +144,15 @@ class readmarker extends forumutils
 	 * @author Geoffrey Dunn <geoff@warmage.com>
 	 * @since 1.3.0
 	 **/
-	function mark_forum_read($forum_id, $time)
+	public function mark_forum_read( $forum_id, $time )
 	{
-		$query = $this->db->query("SELECT topic_id, topic_edited FROM %ptopics
-			WHERE topic_edited > %d AND topic_forum = %d",
-			$this->last_read_all, $forum_id);
-		
-		while ($row = $this->db->nqfetch($query)) {
-			$this->mark_topic_read($row['topic_id'], $time);
+		$query = $this->db->query( "SELECT topic_id, topic_edited FROM %ptopics	WHERE topic_edited > %d AND topic_forum = %d", $this->last_read_all, $forum_id );
+
+		while( $row = $this->db->nqfetch( $query ) ) {
+			$this->mark_topic_read( $row['topic_id'], $time );
 		}
 	}
-	
+
 	/**
 	 * Mark a topic as read to the time set
 	 *
@@ -161,26 +161,24 @@ class readmarker extends forumutils
 	 * @author Geoffrey Dunn <geoff@warmage.com>
 	 * @since 1.2
 	 **/
-	function mark_topic_read($topic_id, $time)
+	public function mark_topic_read( $topic_id, $time )
 	{
-		$topic_id = intval($topic_id);
-		$time = intval($time);
-		
-		if ($time >= $this->last_read_all) {
-			if ($this->guest_mode) {
-				if (!isset($_SESSION[$this->sets['cookie_prefix'] . 'read_topics'][$topic_id]) ||
-						$_SESSION[$this->sets['cookie_prefix'] . 'read_topics'][$topic_id] < $time) {
-					$_SESSION[$this->sets['cookie_prefix'] . 'read_topics'][$topic_id] = $time;
+		$topic_id = intval( $topic_id );
+		$time = intval( $time );
+
+		if( $time >= $this->last_read_all ) {
+			if( $this->guest_mode ) {
+				if( !isset( $_SESSION[$this->sets['cookie_prefix'] . 'read_topics'][$topic_id] ) || $_SESSION[$this->sets['cookie_prefix'] . 'read_topics'][$topic_id] < $time ) {
+					$_SESSION[ $this->sets['cookie_prefix'] . 'read_topics'][$topic_id] = $time;
 				}
 			} else {
 				$this->_load_readmarkers();
-				if (!isset($this->readmarkers[$topic_id]) || $this->readmarkers[$topic_id] < $time) {
-					$this->db->query("REPLACE INTO %preadmarks
-						(readmark_user, readmark_topic, readmark_lastread)
-						VALUES (%d, %d, %d)", $this->user_id, $topic_id, $time);
+
+				if( !isset( $this->readmarkers[$topic_id] ) || $this->readmarkers[$topic_id] < $time ) {
+					$this->db->query( "REPLACE INTO %preadmarks (readmark_user, readmark_topic, readmark_lastread) VALUES (%d, %d, %d)", $this->user_id, $topic_id, $time );
 					$this->readmarkers[$topic_id] = $time;
 				}
-				
+
 				// There is a chance of running a cleanup
 				$this->cleanupchance = true;
 			}
@@ -195,20 +193,20 @@ class readmarker extends forumutils
 	 * @since 1.2.0
 	 * @return date the topics was last read or all topics were read
 	 **/
-	function topic_last_read($topic_id)
+	public function topic_last_read( $topic_id )
 	{
-		$topic_id = intval($topic_id);
-		
+		$topic_id = intval( $topic_id );
+
 		$last_post_time = $this->last_read_all;
-		
-		if ($this->guest_mode) {
-			if (isset($_SESSION[$this->sets['cookie_prefix'] . 'read_topics'][$topic_id]) &&
-					$_SESSION[$this->sets['cookie_prefix'] . 'read_topics'][$topic_id] > $last_post_time) {
+
+		if( $this->guest_mode ) {
+			if( isset( $_SESSION[$this->sets['cookie_prefix'] . 'read_topics'][$topic_id] ) && $_SESSION[$this->sets['cookie_prefix'] . 'read_topics'][$topic_id] > $last_post_time ) {
 				$last_post_time = $_SESSION[$this->sets['cookie_prefix'] . 'read_topics'][$topic_id];
 			}
 		} else {
 			$this->_load_readmarkers();
-			if (isset($this->readmarkers[$topic_id]) && $this->readmarkers[$topic_id] > $last_post_time) {
+
+			if( isset( $this->readmarkers[$topic_id] ) && $this->readmarkers[$topic_id] > $last_post_time ) {
 				$last_post_time = $this->readmarkers[$topic_id];
 			}
 		}
@@ -224,21 +222,22 @@ class readmarker extends forumutils
 	 * @since 1.2
 	 * @return bool True if all posts within have been marked read
 	 **/
-	function is_topic_read($topic_id, $last_post_time)
+	public function is_topic_read( $topic_id, $last_post_time )
 	{
-		$topic_id = intval($topic_id);
-		$last_post_time = intval($last_post_time);
+		$topic_id = intval( $topic_id );
+		$last_post_time = intval( $last_post_time );
 
-		if ($last_post_time < $this->last_read_all) return true;
-		
-		if ($this->guest_mode) {
-			if (isset($_SESSION[$this->sets['cookie_prefix'] . 'read_topics'][$topic_id]) &&
-					$_SESSION[$this->sets['cookie_prefix'] . 'read_topics'][$topic_id] >= $last_post_time) {
+		if( $last_post_time < $this->last_read_all )
+			return true;
+
+		if( $this->guest_mode ) {
+			if( isset( $_SESSION[$this->sets['cookie_prefix'] . 'read_topics'][$topic_id] ) && $_SESSION[$this->sets['cookie_prefix'] . 'read_topics'][$topic_id] >= $last_post_time ) {
 				return true;
 			}
 		} else {
 			$this->_load_readmarkers();
-			if (isset($this->readmarkers[$topic_id]) && $this->readmarkers[$topic_id] >= $last_post_time) {
+
+			if( isset( $this->readmarkers[$topic_id] ) && $this->readmarkers[$topic_id] >= $last_post_time ) {
 				return true;
 			}
 		}
@@ -252,13 +251,13 @@ class readmarker extends forumutils
 	 * @param int $last_post_time Time of newest post in forum
 	 * @author Geoffrey Dunn <geoff@warmage.com>
 	 * @since 1.2
-	 * @return bool True if post has been marked read 
+	 * @return bool True if post has been marked read
 	 **/
-	function is_post_read($topic_id, $post_time)
+	public function is_post_read( $topic_id, $post_time )
 	{
-		return $this->is_topic_read($topic_id, $post_time);
+		return $this->is_topic_read ($topic_id, $post_time );
 	}
-	
+
 	/**
 	 * Check if there are unread posts in the forum
 	 * This is a tricky one because we may have to query the topics in the forum
@@ -267,33 +266,34 @@ class readmarker extends forumutils
 	 * @param int $last_post_time Time of newest post in forum
 	 * @author Geoffrey Dunn <geoff@warmage.com>
 	 * @since 1.2
-	 * @return bool True if all topics within have been marked read 
+	 * @return bool True if all topics within have been marked read
 	 **/
-	function is_forum_read($forum_id, $last_post_time)
+	public function is_forum_read( $forum_id, $last_post_time )
 	{
-		$forum_id = intval($forum_id);
-		$last_post_time = intval($last_post_time);
-		
-		if ($last_post_time < $this->last_read_all) return true;
-		
-		$this->_load_forum_topics();
+		$forum_id = intval( $forum_id );
+		$last_post_time = intval( $last_post_time );
 
-		if (isset($this->forumtopics[$forum_id]) && (0 != $this->forumtopics[$forum_id]))
-			return false;
-
-		$forums = $this->forum_array($this->forum_grab(), $forum_id);
-		if (!$forums)
+		if( $last_post_time < $this->last_read_all )
 			return true;
 
-		foreach ($forums as $f)
+		$this->_load_forum_topics();
+
+		if( isset( $this->forumtopics[$forum_id] ) && ( 0 != $this->forumtopics[$forum_id] ) )
+			return false;
+
+		$forums = $this->forum_array( $this->forum_grab(), $forum_id );
+		if( !$forums )
+			return true;
+
+		foreach( $forums as $f )
 		{
-			if (!$this->is_forum_read($f['forum_id'], $last_post_time)) {
+			if( !$this->is_forum_read( $f['forum_id'], $last_post_time ) ) {
 				return false;
 			}
 		}
 		return true;
 	}
-	
+
 	/**
 	 * Check if there are unread posts in the forum
 	 * This is a tricky one because we may have to query the topics in the forum
@@ -304,45 +304,46 @@ class readmarker extends forumutils
 	 * @since 1.2
 	 * @return bool True if all topics within have been marked read 
 	 **/
-	function _load_readmarkers()
+	private function _load_readmarkers()
 	{
-		if (!$this->readmarkers_loaded) {
+		if( !$this->readmarkers_loaded ) {
 			$this->readmarkers = array();
-			$query = $this->db->query("SELECT * FROM %preadmarks WHERE readmark_user=%d", $this->user_id);
-			while ($mark = $this->db->nqfetch($query))
+
+			$query = $this->db->query( "SELECT * FROM %preadmarks WHERE readmark_user=%d", $this->user_id );
+
+			while( $mark = $this->db->nqfetch( $query ) )
 			{
 				$this->readmarkers[$mark['readmark_topic']] = $mark['readmark_lastread'];
 			}
 			$this->readmarkers_loaded = TRUE;
 		}
 	}
-	
+
 	/**
 	 * Loads in a list of topic ids for a forum so we can check if the topics have been read
-	 * 
+	 *
 	 * PRIVATE
 	 *
 	 * @author Geoffrey Dunn <geoff@warmage.com>
 	 * @since 1.2
 	 * @version 1.3.0 - Altered by Matt L to use fewer querys and removed unused parameter $forum_id.
 	 **/
-	function _load_forum_topics()
+	private function _load_forum_topics()
 	{
-		if (!$this->forum_topics_loaded)
+		if( !$this->forum_topics_loaded )
 		{
 			/* find all topics since we pressed mark all read */
-			$query = $this->db->query("SELECT topic_id, topic_edited, topic_forum FROM %ptopics
-			   WHERE topic_edited > %d", $this->last_read_all);
+			$query = $this->db->query( "SELECT topic_id, topic_edited, topic_forum FROM %ptopics WHERE topic_edited > %d", $this->last_read_all );
 
 			/* read all the records*/
-			while ($row = $this->db->nqfetch($query))
+			while( $row = $this->db->nqfetch( $query ) )
 			{
 				/* Set to 0 if not set */
-				if(!isset($this->forumtopics[$row['topic_forum']]))
+				if( !isset( $this->forumtopics[$row['topic_forum']] ) )
 					$this->forumtopics[$row['topic_forum']] = 0;
 
 				/* increase un-read count */
-				if (!$this->is_topic_read($row['topic_id'], $row['topic_edited']))
+				if( !$this->is_topic_read($row['topic_id'], $row['topic_edited'] ) )
 				{
 					$this->forumtopics[$row['topic_forum']]++;
 				}
@@ -350,7 +351,7 @@ class readmarker extends forumutils
 			$this->forum_topics_loaded = true;
 		}
 	}
-	
+
 	/**
 	 * Deletes unneeded records frop readmarks because they are for topics
 	 * that are older than the last time we've hit Mark All
@@ -360,23 +361,24 @@ class readmarker extends forumutils
 	 * @author Geoffrey Dunn <geoff@warmage.com>
 	 * @since 1.2
 	 **/
-	function _cleanup_readmarks()
+	private function _cleanup_readmarks()
 	{
 		$readable_forums = $this->create_forum_permissions_string();
+
 		// Find the OLDEST unread post
-		$query = $this->db->query("SELECT topic_id, topic_edited FROM %ptopics
-			WHERE topic_edited > %d AND topic_forum IN (%s)",
-			$this->last_read_all, $readable_forums);
-		
+		$query = $this->db->query( "SELECT topic_id, topic_edited FROM %ptopics WHERE topic_edited > %d AND topic_forum IN (%s)", $this->last_read_all, $readable_forums );
+
 		$oldest_time = $this->time;
-		while ($row = $this->db->nqfetch($query))
+
+		while( $row = $this->db->nqfetch( $query ) )
 		{
-			if ($row['topic_edited'] < $oldest_time && !$this->is_topic_read($row['topic_id'], $row['topic_edited']))
+			if( $row['topic_edited'] < $oldest_time && !$this->is_topic_read( $row['topic_id'], $row['topic_edited'] ) )
 			{
 				$oldest_time = $row['topic_edited'];
 			}
 		}
-		$this->mark_all_read($oldest_time - 1);
+
+		$this->mark_all_read( $oldest_time - 1 );
 	}
 }
 ?>
